+++ Changelog.md
@@ -43,6 +43,8 @@
   ([#380](https://github.com/tafia/quick-xml/pull/380))
 - test: added bunch of tests for sequences deserialization
   ([#380](https://github.com/tafia/quick-xml/pull/380))
+- fix: allow to have an ordinary elements together with a `$value` field
+  ([#380](https://github.com/tafia/quick-xml/pull/380))
 
 ## 0.23.0-alpha3